(hds-2069) Fix renamed PW helpers after rebase
