+++ e2e/tests/core/components/core-breadcrumb-spec.ts
@@ -1,6 +1,5 @@
 import { test } from '@playwright/test';
-import { takeAllStorySnapshots } from '../../../helpers';
-
+import { takeAllStorySnapshots } from '../../../utils/playwright.util';
 const componentName = 'breadcrumb';
 const storybook = 'core';
 const takeStateSnapshots = true;
