+++ e2e/tests/core/components/core-text-input-spec.ts
@@ -1,5 +1,5 @@
 import { test } from '@playwright/test';
-import { takeAllStorySnapshots } from '../../../helpers';
+import { takeAllStorySnapshots } from '../../../utils/playwright.util';
 
 const componentName = 'text-input';
 const storybook = 'core';
