+++ e2e/tests/react/components/react-file-input-spec.ts
@@ -1,5 +1,5 @@
 import { test } from '@playwright/test';
-import { takeAllStorySnapshots } from '../../../helpers';
+import { takeAllStorySnapshots } from '../../../utils/playwright.util';
 
 const componentName = 'fileinput';
 const storybook = 'react';
